+++ src/Position.java
@@ -1,24 +1,24 @@
 public class Position {
-  private int x, y;
+  private float x, y;
 
   Position(int x, int y) {
     this.x = x;
     this.y = y;
   }
 
-  public int getX() {
+  public float getX() {
     return x;
   }
 
-  public void setX(int x) {
+  public void setX(float x) {
     this.x = x;
   }
 
-  public int getY() {
+  public float getY() {
     return y;
   }
 
-  public void setY(int y) {
+  public void setY(float y) {
     this.y = y;
   }
 }
